Enabling healthchecks for native LBaaS driver leads to vrouter crash
The crash is happening because we are asserting on exception while executing
db_table task. The exception is happening because while calling substr we are
passing offset value greater than the length of the string which throws
out_of_range exception.

To fix this, we need to check offset length should not be greater than the
length of string.

closes-jira-bug: CEM-9813
Change-Id: I906b65608e57e6a53080261b5c8b0f0edde346c4