Prevent route adds to delete-marked VRFs
AddEvpnEcmpRoute does not check to see if the VRF has been marked for deletion
before going ahead with addition. This resulted in AddEvpnEcmpRoute accessing rt
table's vrf_entry_ which was NULL. Added check at the beginning of
AddEvpnEcmpRoute to bail out if VRF has been marked for deletion. Same change is
added to AddEvpnRoute.

Closes-Jira-Bug: CEM-10132
Change-Id: Ic818b021ead3bbd868479b6de566ac21d1cb7e48