K8s/Contrail - vRouter issue when sending TCP Segments back to Compute with higher MTU
When a pkt having size > outgoing interface MTU is received by vrouter, it traps the agent
to send "ICMP fragmentation required" msg to the sender.
In case of NAT flows, this msg was not being sent due to errorenous check.
Fixed it to check for NAT flows correctly.

Fix has been verified by submitter.
Agent gtest UT is not possible for this case as there is no support
to check for sent pkts.

closes-jira-bug: CEM-9393
Change-Id: I2e61a15807ea3847111a10cf78e8986b17884a3c