L2 route on vrouter is pointing to incorrect compute host
Ksync is getting 2 request with same nh id but different dip and hence it
is creating 2 ksync nh entries. The 2nd ksync nh entry may be because of a
corner case where a nexthop db entry which has been marked deleted
but we went ahead and created ksync nexthop entry for that nh since
there is no check for mark deleted case. Hence included the check.

closes-jira-bug: CEM-10272
Change-Id: I38c64de16f2ad3c30b792d5512456c6f91ee11d0