Add support to Connection timeout in vnc api lib closes-jira-ticket: CEM-10694
As part of CEM-8756, timeout config parameter has been added to
vnc_api lib with a default value of 5 secs.
However this timeout was applied to both connection as well as
request timeout. This affects if a response from API server was not
received in default 5 secs.
This commit introduces a new parameter called connection_timeout
and let use timeout introuduced in CEM-8756 as dedicated
api server's request timeout

Change-Id: I83a3d223281e254e363b2264174e60769e98055b