Do not send RA for BAREMETAL VMIs if the VN belongs to logical router
Currently there is no check to prevent RA being sent if the baremetal VMI
belongs to a VN that is associated with logical router. The RA funcitonality
will be taken care by the connected qfx.
Adding the check to prevent RA being sent.
UT infra for Baremetal VMI enhancement is tracked via cem-10960

Closes Jira Bug: CEM-10885

Change-Id: I05aee8011d11fc643e0fbe27b4b8ff3813b35ec9