Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4: Code-Review-1
> > 
> > (2 comments)
> > 
> > You don't validate anymore the LR route target field?
> > You should create a generic method to validate resource RT fields and define a constant map in the GSC class to know on which resource's fields need to check. Then in the future, we will just have to update that map if new RT field are remove/added to the schema.
> 
> Yes I dropped LR validation. It must be discussed because existing UT for LR fail when new validation is enabled.

This is not normal. I think that UT are wrong so. Do you prefer to address that in two patches, one for VN and another one for LR? If yes please put that on as partial for the bug. Also can you consider my review to propose a generic method to validate resource fields which contains user defined RT in the first or the second patch?

Patch-set: 4
Reviewer: Gerrit User 10 <10@c72cd27d-1743-4a55-84a4-cd13b8400c52>