Fix VxlanRoutingEnabled unit test
Issue:
When the Peer order changed from EVPN_ROUTING_PEER to BGP_PEER,
the local route active path type will be “interface NH” not the “vrfNH”
.
Local routes in the INET table with EVPN_ROUTING_PEER are the routes that are populated by the Agent derived from the EVPN-type2 routes. For these derived routes NH is of type "VrfTranslate" (VrfNH). Local routes in the INET table have following paths (when Vxlan routing is enabled):
1) derived routes (from EVPN-type2), Type VrfNH.
2) BGP peer advertised, Type InterfaceNH
Due to the order change, the active path will be BGP peer advertised.

Since the routing the unit test V2_VxlanRoutingEnabled is specifically checking forthe active path of type vrfNH so was the test failing.

Fix:
The interface type that need to be checked is of type interface NH.

Cloes-Jira-Bug: CEM-10466
Change-Id: I98919c900f4213a2d0329f324684ac69d3279bac