[config] Remove pycassa dependency to run UT with py3
As contrail-config-common package depends on pycassa and as pycassa does
not supports python 3, we were not able to run any config service UT.
That patch removes pycassa dependency everywhere and replaces it by the UT
mock. That will be remove when we will get ride of pycassa in python 2
(CEM-8384).

Change-Id: Idc76ec618fd4e1f901fae464e25398db0edd63cf
Closes-Jira-Bug: CEM-11521