Fix EVPN Transparent service-chain
Issue:
For a transparent service-chain type, the lookup for the tail-end of
the service instance will be done in the service-instance RI
and not in the VN-RI associated with the service VM. This is due to
vrf-assign rule policy in Agent.
In the EVPN vnf service chaining, traffic egressing the tail-end of
the VNF doesn't have the type5 route for the local path since Agent
only populates EVPN Type5 routes in routing Vrf and not in the service-chain vrf.

Fix:
In EVPN VNF service chaining, control-node sends the route originated
by Agent in routing-vrf for the service-chain vrf as well.
When encapsulation used is VXLAN, nexthop cannot be found from the
message. Since EVPN Type5 local-route is not available in the
serivce-chain vrf evpn table, set the next-hop for the local route
(in the serice-chain's vrf) to primary-routing-vrf to have the common
design for the EVPN Type5.

Tested following cases:
a) EVPN Transparent service Instance across different computes
b) Transparent ECMP case
c) Innet SI type, service instance spawned across different computes
d) Innet SI + Transparent ECMP SI

Closes-Jira-Bug: CEM-10608
Change-Id: Icc6b17393358a2b26ea7ba98d6fde80d5ae9a571