Update patch set 34
Patch Set 34:

(12 comments)

> Patch Set 33: Code-Review-1
> 
> (12 comments)
> 
> Still not clear to me why you add hooks in the resource code to update the PR annotation. Why the code which is responsible to create/update resources related to that fabric job/transaction takes care to update that annotation? It looks we complicates the core API code for specific fabric task.


The goal is to correlate the intent transaction (operation) in the UI to the resulting config push job(s) in the device-manager. For multi-device operations, there will be at least one device config push job per device and we want to tag all these jobs with the same transaction ID. Therefore, on the job status page in the UI, we can group all jobs of the same transaction onto one expandable line.

Device-manager generates and pushes device config based upon triggers from VNC DB updates and a reaction map which reacts to anything that requires new config to be pushed to a device. This reaction map is implemented on a per-device basis, so the frame of reference is a single device. In order to implement transaction logic in device-manager, it becomes very difficult, in some cases, to trace a particular VNC DB operation back to the original UI intent without knowing the full context. Plus, we want to use the same transaction ID for all devices affected by the intent operation so there becomes the added challenge of correlating the device jobs together. Finally, there is additional overhead in device-manager reading VNC DB objects again, when api-server has already done so in the resources logic.

Patch-set: 34
Reviewer: Gerrit User 510 <510@c72cd27d-1743-4a55-84a4-cd13b8400c52>