Update patch set 39
Patch Set 39:

Hi Joe,

There is a time delay between when a transaction-id/descr update is made to the PRouter object and when DM reads the same within which if someone edits another object then the PRouter's transaction info would get modified and the Job would be wrongly mapped to the latest event.

This is a problem since API-DM communication is asynchronous.
Believe we could overcome this limitation if we could send the transaction info as part of RMQ update or by generating transaction info in DM itself.

Let me know if my understanding is wrong.

Thanks,
Senthil

Patch-set: 39
CC: Gerrit User 126 <126@c72cd27d-1743-4a55-84a4-cd13b8400c52>