UT added for CEM-11283
The change in trace buffer capacity is correctly updated in the trace
container. But it is not udpated in a variable in TraceBuffer object.
Applications use this to query the size before request for the trace
contents. Fixed this issue by updating the information in TraceBuffer
object in addition to actually updating the Trace contrainer capacity.
UT added

Closes Jira Bug: CEM-11283

Change-Id: I138a226e8d90fbd4aaf2aa8e7edd58cdb2b2896d