[Device manager] Add another method of reloading dnsmasq
It is assumed that device manager and dnsmasq services are deployed always as docker containers on node that has docker installed. It is not always true. In case of kubernetes deployments a different runtime might be used like containerd.

Since device and dnsmasq containers are strongly coupled they will always run in a single POD and share the same process namespace. In that case device manager can just restart dnsmasq process directly.

Closes-Jira-Bug: CEM-12148
Change-Id: Idbc6825c90af859928fb82516d6237e3e63f4e2c