Fix EVPN Type5 Routes not getting deleted in service-instance VRF
Issue:
For a transparent service-chain having a local route is getting added with
different key and when the retract for the same route from the control
received was trying to delete with different key.

Fix:
Using same key for adding and deletion of the Type5 route in service-instance's
vrf.

Tests:
i) Ran automation test script "TestSVCCombosForEvpn.test_evpn_svc_combo_all"
with transparent service chain enabled.
ii) Added UT for validating the same.

Closes-Jira-Bug: CEM-11952
Change-Id: I541dca9223de7eb5f9637381d9bf973779be68f8