[config] Sub-cluster ID allocation
Allocate a unique ID for each Sub-Cluster. A user can define the ID, API
will verify if it is free or not. If the user does not provide it, it
will allocate one.
If the global ASN is moved from 2 to 4 bytes, code takes Sub-cluster ID
can fit in the reduced space.

TODO: allow user to change the Sub-Cluster ID

Closes-JiraBug: CEM-11272
Depends-On: Ifb58392b671135f6d32d3a095a7c2829730e4f8b
Change-Id: Ib2cedaf4d563c6e2fc927d79ddd0601d689e886d