Fixed LPM issue in ip-fabric-forwarding
Issue:
If the workload from ip-fabric-forwarding's VRF gets the floating ip
from the same VRF and tries to reach other workload on the same VRF,
floating ip is used as the source address instead of the fixed ip for
the same prefix since floating ip is preferred over fixed ip. The
issue is not seen in the overlay VRF. It is because there is an
additional check avoids snat and uses the fixed ip. The check compares
out->vrf->rt_ and floating_ip->vrf->rt_. since the rt is same, fixed ip
is used. In ip-fabric-forwarding VRF the additional check fails since
out->vrf is diffrent from floating_ip->vrf. so floating_ip is used as
the source address

Fix:
Ideally the fixed ip should be used than the floating ip when route
prefix of fixed ip >= route prefix of floating ip. The fix is doing
the same.

Closes-jira-bug: CEM-12277
Change-Id: I3165e6449049cdf6fdba70bb41d2fa2488a7cda8