Update patch set 5
Patch Set 5:

(2 comments)

Please do not fix pep8 errors in the file to not pollute the commit logs. Hope in the future we will be pep8 compliant but that needs to be addressed in on or few dedicated commits.

Patch-set: 5
Reviewer: Gerrit User 10 <10@c72cd27d-1743-4a55-84a4-cd13b8400c52>