This will support ip allocation for routed_vn virtual network
Currently contrail addr mgment supports flat subnet and user-defined
subnet. This feature is only applicable user-defined subnet which is
defined in the link between ipam and virtual network.
contrails reserves 4 ip addresses in any subnet, first, last,
first+1(gw-ip) and first+2(service-ip). If virtual network is configured
as a routed vn, first+1 and first+2 will not reserved for gw-ip and
service-ip. These two ip addresses will be available for instance-ips as
long as Subnet is not flat and defined as a user-defined subnet

Change-Id: I5b6c09f7bef48a2d7cc8134f0ef860936a30dfe6