Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Can you add version number similar to https://review.opencontrail.org/c/Juniper/contrail-controller/+/57199/1/src/config/api-server/vnc_cfg_api_server/db_manage.py . Use 27 as
> 26 already being used in this review.

I think it is not a good practice to skip 26 and go straight to 27 because we shouldn't assume the merge order. But I did as you asked me. I'll resolve the conflicts when your PR will be merged. Let me know when it merges because it's easy to miss an email from zuul.

Patch-set: 1