Fix EVPN route add crash in VNF service-chaining
issue:
CN send the EVPN route for the local route that was advertised
by the same compute node for the service-chain vrf. In service-chaining
configured between the LR's (IVN), Agent allows local routes for EVPN-type5
in the service-chain vrf.
For service-chaining between VNs (not the IVN), the code was deriving
the logical-router
vrf, there by it is crashing since logical-router is not applicable for VNF
service-chaining.

fix:
Check if the service-chain vrf's primary vrf is of type routing-vrf (Logical router)
unit test was added to verify for remote and local EVPN type2 route behaviour
for the service-chain vrf.

testing:
1) unit test "service_chain_evpn_local_remote_route" is added.
2) ran automation scripts.svc_firewall.test_evpn_svc_combos.TestSVCCombosForEvpn.test_evpn_svc_combo_1
   and scripts.svc_firewall.test_svc_combos.TestSVCCombos.test_svc_combo_1

Closes-Jira-Bug: CEM-12892
Change-Id: Icf1fc5816764037432a88227aca46d1c127133bc