Add NULL check for component nh key member
when one of teh component member is deleted , we reset that member
position with NULL to support flow stickiness.
in this scenario, route update for default VRF route accessed
NULL member and caused crash,

this code flow is missing check for NULL component NH member of composite NH,
added check for the same.

Closes-Jira-Bug: CEM-13272
Change-Id: Icec1b5a08234ddcc8a2e5856cc5e4935a5b8b36d