Changes to register with GMP only when IGMP enabled.
Register IPAM with GMP only there is an interface
with IGMP enabled on it.  First VMI to be enabled
with IGMP triggers IPAM registration with GMP. Last
VMI to disabled IGMP in IPAM will de-register from
the GMP.

Also, using VMI list from multicast bcast group
as local olist to send query messages instead of
relying on route lookup.

Reorganised and added UT for the same.

closes-jira-bug: CEM-13896

Change-Id: Ie4a58c479f8e50fef454f02a9b69035387949ea5