Fixed vm<->virtual-router link issue in _sync_pod_vm()
_sync_pod_vm() will link vm with virtual-router, if the
virtual-router link is missing. Since it is not using the
right prerequisites for the validatation, the linkage
would not happen for the valid cases too. Modified code
to create the linkage for the valid cases.

closes-jira-bug: CEM-13895
Change-Id: I90b24c28f6f07a3564fe6665137e090d2f1ac7fd
(cherry picked from commit 5ea276056bfe8f61b1b4fc5bb5f1ccf166e9c213)