config: fix sources directory for noarch targets
In this commit we are setting absolute path to refer the SConscript
related to noarch projets. It looks that is fixing issues related to
the sources directory that still refers to #/build/debug.

We also set a new env variable TOP_NOARCH which will be used in futur
to refer #/build/noarch directory.

Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>
Change-Id: I371030b092ee86ff85224d7233c515d2c0275990