R4.1.4.x: Fix BGPaaS session not coming up on VNF reboot in scaled setup
issue:
In a scaled setup with about 2000 BGPaaS session and 1000 BFD sessions,
during VNF start and stop, some of the BGPaaS sessions are down.
The error handling for the ksync flow message is deleting the BGPaaS
sessions flows when received EEXIST error.

fix:
If received EEXIST for BGPaaS flows, don't make the flows short flow
treating similar on par as regular flow processing.

testing:
The fix was tested at customers scaled setup and din't see issues
when EEXIST was noticed.

Closes-Jira-Bug: CEM-10354
Change-Id: Id5ac461f1bdeb94dde719f04e472e81c1b27a66c