config/common: use a mutex to avoid pressure with get_free_port()
To avoid pressure during concurrent tests running and to avoid
returning ports that are already in use. In this commit add wrapper
with a mutex to the method responsable to find and return free ports.

We are expecting with this change to resolve the related issues
"connection refused" during unit tests execution.

Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>
Change-Id: If1526b4770f5c064f0efb60b624a2c2cae51a36f