Check for IPAM NULL.
The check for IPAM equal to NULL is needed at least because of the one
scenario seen during testing.

1. Launch 2 VN
2. Launch one vm in net1
3. Verify vm in contrail components
4. Change the interface of vm to net2

During above scenario, primary_ip_addr_ of VMI takes time to get
updated and causes IPAM query to fail. Adding a safety check for
the same.

Also, restoring the igmp enabled check to before.

closes-jira-bug: CEM-13963
Change-Id: Id8d00091ff64a6afdec5f3d2b8ac6b55b00e0513