Add ref from FMG label to NH
When multiple BUM trees are received one after another varying in FMG label, it is possible that NH ADD and DEL gets processed before MPLS label add/change nottification is processed by ksync module. This results in NH getting freed and the freed NH being used in initializing mpls ksync entry where it creates an NH ksync entry using the  freed NH. This casuses a crash in mpls ksync processing of label change or delete

Assume tree1 is received causing label L1 and path P1 to point to NH1. If, say tree2 is receied with label L2 requiring P1 and L2 to point to NH2. Before label L2 change is processed by mpls sync, following events happen - 1) New tree with label L3 is received requiring L3 and P1 to point to NH3, 2) P1 change causes NH2 to be deleted, 3) NH2 add is either seen or not seen yet by NH ksync. In both cases NH2 is freed (note that the NH pointer MPLS label has is not a ref). When change is processed by ksync MPLS for L2, it ends up using freed NH.

Solution is to add a ref from MPLS label to NH. This is not possible as NH already has a ref to MPLS label. It would result in circular dependency. However the NH to label ref is not used in case of FMG labels. FMG labels already have a map map fmg_nh_list_ to store the VRF and NH they have been assigned to. The NH pointer in fmg_nh_list_ is converted in this fix to ref to ensure that NHs don't get freed until mpls ksync has seen the label change.

Closes-Jira-Bug: CEM-14317
Change-Id: I38a90279465a7dc73a85b2f8b2a33e85333f7995