Fix for mock ports to be active in mock agent
Partial-Jira-Bug: CEM-12108

The earlier code related to this change is
Change-Id: I60547736be3c97399df686441b5707b1688b5151
in controller repository.

BuildVrfAndServiceVlanInfo that gets called from
InterfaceTable::VmiProcessConfig does not have the
direction value set in the mock mode. Because
of this the processing is halted and vrf_name_, ipv4_active, etc.
are not set - making the port inactive. Also the vrouter-port-control
change is required to print the request.

Change-Id: I93b53280338a56584c518fa5d28a3c0a1e75c503