CEM-14660 - Avoid using /dev/vr_bridge and /dev/flow devices in agent             when kernel is using hugepages
On newer ubuntu kernels, a traceback related to bad PTE entries is seen
when agent is exiting. The same is also seen when flow and rt commands are used.
The traceback is seen when unmap for /dev/vr_bridge and /dev/flow is being done.
It is not clear what is causing this. The same is not seen on older kernels used
by RHEL and Centos.

To avoid this issue, agent/flow/rt need not use /dev/vr_bridge and flow devices
since in kernel with hugepages mapped, agent/flow/rt can directly use the mmaped
/dev/hugepages/bridge and /dev/hugepages/flow files.

Agent changes:
1. Send the mmapped hugepage file path to vrouter at the time of init.
    The path is given by vrouter to flow and rt commands.
2. Use the mmapped hugepage file path for accessing flow and bridge table
   instead of /dev/flow and /dev/vr_flow devices.

Testing:
1. Tested with kernel mode without hugepages enabled
2. Tested with kernel (RHEL/Centos and Ubuntu) with hugepages enabled
3. Tested with DPDK compute

closes-jira-bug: CEM-14460
Depends-On: I2a916c23088221d8a2ee236bac3a4bde78384142
Change-Id: Ifa93bc7a6963f230b764e470ef4192bec62e0652