[DM] Fixing brownfield onboarding failure in non-openstack deployments
During brownfield onboarding, in a K8 or a vCenter deployments, at present
we try to create a default-LR under admin project. However there is no
concept of admin project in non-openstack deployments. This causes the
brownfield fabric discvoery workflow to fail.

To fix this problem, we are bringing in the concept of default-project
which is contrail notion and is therefore cluster independent. Also, we create
this default LR not during onboarding but during cluster upgrade or cluster
re-install.

Change-Id: Ie1868f45f9d772e1818a9722f841f15e705bcf98
closes-jira-bug: CEM-14400
target-release: R2005