Update patch set 9
Patch Set 9: Code-Review-1

(2 comments)

Sounds much more better to me to use a ref to keeping trace of the resource generated by the system. This is easier to identify them for deletion or decide actions. But I don't understand why the ref to the loopback VN is from the Fabric resource and not from the FN (the resource which creates it)?

Patch-set: 9
Reviewer: Gerrit User 10 <10@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Label: Code-Review=-1