config/cassandra: collecting profiling statistics of Cassandra calls
In our effort to build a driver that supports Python3 we also want to
improve the current performances of our Cassandra model and
implementation.

In this commit, we add support of profiling Cassandra calls from the
driver perspectives.

Using `CONTRAIL_PROFILE_CASSANDRA=1` before starting the service will
enable the feature. It is to note that collecting calls statistics is
CPU time-consuming as IO consuming. We should not use this feature in
production or during a reasonable amount of time.

The statistics collected will be stored in files. The results may be
read using the `pstats` modules or with software like kcachegrind to
help finding bottlenecks.

CONTRAIL_PROFILE_CASSANDRA=1 python -m unittest vnc_cfg_api_server.tests.test_vnc_cfg_api_server.TestVncCfgApiServer

Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>
Change-Id: Iab0231f6cb22404e7cb5a37dfbdc061d9b0885fb