[API] VN (routed/tenant) can not belong to multiple LR's
Except the case when the vn is a shared VN, a VN cannot belong
to more than one LR in the same fabric at anytime. Implementing
this check on api-server based on approach for performance optimization.

Modified code for factoring in fabric uuid based LR check.

Modified code for factoring in the following cases:
1. Fabric update / delete as part of LR update
2. VN update / delete as part of LR update
3. Finding fabric info based on PR if not directly available

Todo: Modify UTs based on this new check.

Change-Id: Iacfea1fe0ac05ec7740e0924fb46f21b2b1558a3
Depends-On: I688c09a02f794b0a35bad1497170a95e4aee69b4
closes-jira-bug: CEM-12445
target-release: R2008