[API] VN (routed/tenant) can not belong to multiple LR's
Except the case when the vn is a shared VN, a VN cannot belong
to more than one LR in the same fabric at anytime. Implementing
this check on api-server based on approach for performance optimization.

Modified code for factoring in fabric uuid based LR check.

Modified exisitng UTs based on the check for fabric for those
cases where LR had a VN and a VMI.

Todo: Modify UTs based on this new check.

Change-Id: Iacfea1fe0ac05ec7740e0924fb46f21b2b1558a3
Depends-On: Ie649c194eb4cc508c15374bf1aee03f35f687a98
closes-jira-bug: CEM-12445
target-release: R2008