[Fabric] Changing DC-GW irb MAC for MX
When ERB-Ucast role is configured on QFX5110 boxes and when the MAC of the DC-GW
irb happens to be same as that of irb present on QFX5110 box, external ping is
expected to fail from BMS due to QFX5110 limitation. This is addressed by changing
MAC of irb belonging to DC-GW(public VN).

Closes-Jira-Bug: CEM-14045
Change-Id: I0a3608f6b1bcaae7736471a82959ffa6d5970aba