Issue: Control and DNS node crash periodically when ssl is enabled
Root Cause:
The issue happens during the XMPP connection handling of a handshake failure.
The crash happens when try to trace the reason for the failure. This is
because we transition to handle handshake failure that can delete/clear the
XmppConnection object and we try to trace the reason for the failure after
the transition when we end up using a part of the XmppConnection object.
This may not happen all the time since it depends on when the memory is
returned back to the system.

Fix Description:
Trace the handshake failure before transitioning to handling the failure.

Testing:
The fix has been given to submitter for soaking and we have not seen the
crash.

Commit branches:
R2005 and master

Closes-Jira-Bug: CEM-15433
Change-Id: I3f73ae895159f9163302d5e2b7fbfa999b4cd3b8