Set default route 0.0.0.0/0 for virtual network
Normally, when create a VM in a virtual network, we will get a
default route 0.0.0.0/0 with next hop is default gateway ip.

However, in Contrail, setting up 'host_routes' that cause losing
the default route above and cause some problem when cause some problems
with metadata and layer 3 traffic.

This patch is not ideal solution but can fix above issues.

Change-Id: Ib323776e93c831d12a573c400ce197151fbab33f