Clear the call back hooks on stopping instance task. 
When Healthcheck service is stopped, health check instance task is stopped
before posting the STOP_TASK event. This is done to ensure that no more
MESSAGE_READ events are queued by the instance task after STOP_TASK is queued.
Stopping of the task is done by sending SIGTERM to the task. It is possible that
the instance task can queue MESSAGE_READ events after STOP_TASK is queued, but
before it is terminated with the signal.
The fix clears call back functions on stopping the instance so that no more
MESAGE_READ events are queued once the Stop() is executed on the instance task.
This is done under the protection of a mutex to avoid any race condition in
using the call back hook variables.
 
Closes-Jira-Bug: CEM-15367

Change-Id: I78d11f03817f17aa428197954dc0149b538d02da