fix using filter(...) for python3
python2 returns list for this object but python3 returns 'class filter'
so we must convert it to list where list is required

with unblocked installations of py3 UT new issue came -
genventhttpclient==1.0a can't be installed with python3.
pinning geventhttpclient to 1.3.1 will not help cause it
raises an error at packaging stage into containers.
But current code-base doesn't use grventhttpclient at all.
So it can be removed.
geventhttpclient is a new library instead of removed
gevent.httplib. But this import also not in the code.

pin gevent<1.5.0 for most modules
schema-transformer uses old module that produces many errors for UT
and it can conflicts with common modules that uses <1.5.0 restriction

related-jira-bug: CEM-14275

Change-Id: If4a9e1e5e2ae94cd4c729232e57a1ce76d7820de