fix using filter(...) for python3 and fix gevent/geventhttpclient issues
python2 returns list for this object but python3 returns 'class filter'
so we must convert it to list where list is required

with unblocked installations of py3 UT new issue came -
genventhttpclient==1.0a can't be installed with python3.
pinning geventhttpclient to 1.3.1 will not help cause it
raises an error at packaging containers stage.
But current code-base doesn't use geventhttpclient at all.
So it can be removed.
geventhttpclient is a new library instead of removed
gevent.httplib. But this import also not presentin the code.

add workaround to schema-transformer with gevent library.
its UT doesn't work with gevent>1.1.2 for python2
and gevent==1.1.2 can't be used with python3
Apply theworkaround as markers are not supported in requirements.txt

related-jira-bug: CEM-14275

Change-Id: If4a9e1e5e2ae94cd4c729232e57a1ce76d7820de