config/cassandra: CQL driver implementation
From this patch vnc_cassandra will be using the CQL's driver under
py3's env.

It is also possible to force usage of the CQL driver for py2's env
that by initializing the OS ENV variable CONTRAIL_EXP_CQL_SUPPORT to
1.

In this commit we add the full driver implementation, we update the
fake Cassandra server to add a comprehensive support of CQL and
finally an encapsulation to fake direct calls to Cassandra servers.

Closes-Jira-Bug: CEM-10731
Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>
Change-Id: I99dff2fb81c05d8e20ea7f6bfd0e27f422a8eafa