Change route addition for metadata ip to happen in db::Table context
When a BgpaaS flow is created, FlowMgmt task triggers the creation of health
check instance which results in adding a link local route to default vrf. The
route and path creation happen in the context of FlowMgmt task. Addition of the
route to DBTable triggers a notification to listeners. In the time interval
between addition of route and the addition path to it, if DBTable task runs and
the notification is processed by Route Ksync object, it finds the route path
list empty and hence the active path to be NULL resulting in a crash on its
dereferencing.
Note that the FlowMgmt and DBTable tasks are not excluded from running
concurrently. The fix makes FlowMgmt to enqueue the route-add request rather
than adding in line. This ensures that the route is added in db::Table context
and hence the notification handing can't run until the path is added to the
route.
One UT case has been enahnced to simulate the scenario.

Change-Id: Ib19067311fc8cdfd116b3f5b03a3ee7e6cc9946e