Update patch set 14
Patch Set 14:

(3 comments)

I appreciate your review Edouard thank you. Please understand that the aim of the patches is not to change the logic or making refactoring, it's about to put everything related in a clean interface which will help to build the new driver without adding regression, in the opposite of what we currently have which is fragmented in small parts of code everywhere.

Patch-set: 14
Reviewer: Gerrit User 40 <40@c72cd27d-1743-4a55-84a4-cd13b8400c52>