Change agent to not be tied to vhost0 IP and oper state Closes-Jira-Bug: CEM-15411
These changes are needed as part of the effort to obviate the
need to move interface attributes from the physical interface
to vhost0.

We allow the physical interface IP binding to used as the trigger
for the agent to connect to the controller, rather than just the
vhost0 IP binding. It is expected that this is only used for the
kernel datapath. The assumption is that in the case of DPDK, the
IP address is already moved to vhost0 by the time the agent code
runs.
We also don't check for vhost0 oper state of UP before connecting
to the controller in the case of the kernel datapath.

Signed-off-by: Darrell Ball <dball@juniper.net>
Depends-On: Ibbdfc71e5d01ddf5661901056f4ebfbef51a57d4
Change-Id: I94d4424742485e339c37d24f92d715fe6f018e16