another annoying test failure fix
test_public_snat_routes sends a request to create a logical router
instance. the object has a complex structure and it takes much time
to create all of its components. the test now waits for creation
of the service-instance and expects the router is fully functional
when the wait finishes. this is not always the case and leads to
false positive sporadic failures.
let's add another wait for 5 seconds to make sure the router is ok
or throw a reasonable error instead of dumping the stack because of
npe.

Change-Id: If46ed6c4804900ca834ab14a0c3175cda86b60de