use native python member to evict an item from a dict
the approach used so far to test for existance and then del in the
end is not secure because the dict can be updated between the check
and delete. that case shows in tests.
let's leverage python dict functionality to pop an item by key if
the key is in the dict and report None otherwise, then check the
result and proceed in case it is valid.

Change-Id: I94985c7582fdfcbac0811f3d2d24628b3720f373