Update Tunnel NH's ksync entry if id does not match with TNH
It was found that when previous incarnation of TNH still exists when the new
incarnation is created, the new Ksync entry was getting initialized with old
TNH. If the old TNH had not yet been seen by Ksync, it is possible for the new
TNH to get linked with the NH KSync entry intited with old TNH. This resulted in
the new TNH and its ksync entry having different ids. Note that NH Ksync entry
does not use NH id as part of its key. Fix updates the Ksync entry's nh id and
nh pointer when the mismatch is found.

Closes-Jira-Bug: CEM-16354
Change-Id: Icada5ada6ba81842b871d2f31edd6564eea76c1f