Set the max open files to the required value via setrlimit
Set the rlimit for max open files to the required value of
linklocal_system_flows_ + Agent::kMaxOtherOpenFds + 1. Agent was not setting the
soft limit if it is higher than the required value. It is changed to set to the
required value even if the existing soft limit is higher than the required
value. This is done to avoid long close() FD loops following any fork or vfork()
agent may do in case the soft value is set to max allowed value of 1M.
A UT case has been enahnced to draw attention to the issue if
Agent::kMaxOtherOpenFds is modified.

Closes-Jira-Bug: CEM-16332
Change-Id: Iefccf7088d16b6f5f984b209fb6d29b1fed48c4e