Fixed breakage in the host_network_service feature
The actual issue is the backend pods are enabled with hostNetwork.
When we create the link-local, all traffic goes to the fabric ip
and from there it goes to the backend pods via the iptables rules
created by the kube-proxy. "host_network_service" features solves
the issue with vrouter. This feature was added as part of 5.1. By
default it is disabled. If we enable it, the k8s service ip will be
associated to all the backend pods including the pods which has
hostNetwork.

Right now the k8s service ip is not associated to all hostNetwork
backend pods due to the validation check added for the overlay vmi.
The validation check is not needed for the hostNetwork backend pods
since there is only one vmi(vhost0) for each host.

partial-jira-bug: CEM-15133
Change-Id: Ifb66d95e38c3565973734583f15314f21276abe7