support k8s with containerd backend
ubuntu use case. containerd provides cri api thus we can work with
it using an existing implementation. the only difference with the
cri-o scenario is a different path to unix domain socket.
detect if we are running inside a containerd managed container and
connect to its socket to do the job.

partial-jira-bug: CEM-16840
Change-Id: Idd1d5dfa8712bba3f67d3faa68a2017f212343c0