Set pbb_mac for vmi only if pbb_interface flag is set for it.
When interface is added in agent for corresponding ksync interface
isid and pbb mac is set only if pbb_interface flag is set for the
interface. This check is missing in Sync for ksync interface, in
this case pbb mac gets set incorrectly (should not be set) if sync
is called for entry. Add check to set pbb mac if interface has
pbb_interface flag set . Initialize pbb mac.

Add unit test.

Closes-jira-bug: CEM-15162

Change-Id: Ifdd8b5bb34a6ee7fd381461d355d3ffe27c1905d