Ifmap changes for link revival
There were multiple issues found in ifmap link revival i.e.
delete quickly followed by add. In such cases, old link is
reused along with the state. This state will have list of
receivers (advertised set) and an update does not go to any
receiver who is in this advertised set. We need to reset this
advertise set while deleting a link. Also, link needs to be
recreated which will trigger sending down of updates. Unit
test was there but it was expecting wrong thing so correted
that as well.

Change-Id: If3e12c2b07c44f994aae369d0b0b70462f1824dd
Closes-Jira-Bug: CEM-16654