config/common: fix unrelevant stats/latency logs
Only methods of vnc_cassandra should be decorated by VNC stats/latency
and reported to Customer. For specific performance profiling related
to direct calls to Cassandra, please refer to
https://review.opencontrail.org/c/Juniper/contrail-controller/+/58345/17

NOTE: The behavior with reconnection to Cassandra is connection gets
lost stay the same as the exception will be still reported and
handled.

Closes-Jira-Bug: CEM-13898
Closes-Jira-Bug: CEM-16601
Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>
Change-Id: I2a99da6cf6840a5d219e7e237381bfb327460a3f