Update patch set 3
Patch Set 3:

Thanks for your review Nagendra. 

I think that you are bringing a good point. But first, we need to understand use-case of this tool.

It is used as a best-effort backup during a disaster, or is it used to provide consistent backups to customers?

Patch-set: 3
Reviewer: Gerrit User 40 <40@c72cd27d-1743-4a55-84a4-cd13b8400c52>