fixes failing device-manager container
DeviceZtpManager needs to remove every file except for base.conf
from /etc/dnsmasq folder. it lists everything and tries to remove
the name even if the file doesn't exist which leads to an exception
because python list.remove throws in such a case.
let's ignore the case when base.conf file is unavailable originally.

Change-Id: If10aa6207b37987d599797c45339f00b45af1575