Fix to handle "/" in tags in kubemanager
Since kubevirt has "/" in its labels, kubevirt vm traffic
is failing due to the missing tags.
For each label in k8s, custom tag is created in contrail
config and fqname is stored in zookeeper. Since zookeeper
follow the directory structure model, tag with "/" would
not treated as single entity which is causing the issue.
Fix is added in kubemanager to replace "/" with "__slash__"
if tag has "/".

Closes-jira-bug: CEM-18263
Change-Id: I93c7907949e83303a19b81c6aeaa2bb91651e7c6