ZK locks for delete VMI should include VPG or fabric
closes-jira-bug: CEM-18266

Author: Nagendra Maynattamai npchandran@juniper.net

Zookeeper locks should be created for a specific VPG or fabric.
This was done properly for lock used during create/update VMI objects
However for the locks created for delete, this change was missing.

This patch will create locks based on specific VPG or fabric UUID
during delete VMI objects

Change-Id: Id6e4a5804afca6d4fa77bac55445a6770e8784d1