Optimized UVE Interface SG Rule Stats Collection - contrail-vrouter-agent
UVE Interface SG rule stats are collected for every 30 secs. It is
collected for non active sg rule uuid's too. It creates too many
data when system have lot of flows which makes the system unstable
(if the system is capacity is low). Optimized to collect UVE SG Rule
Stats for the active sg rule uuids.

closes-jira-bug: CEM-18194
Change-Id: I619e4746400e2580176047d775bb5c85e9278ff0